// src/app/enums.ts
export enum GameType {
    WORDLE = 'WORDLE',
    DORDLE = 'DORDLE',
    QUORDLE = 'QUORDLE'
}

export enum LineStatus {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE',
    FINISHED = 'FINISHED'
}

export enum ActionType {
    ADD = 'ADD',
    DELETE = 'DELETE',
    SUBMIT = 'SUBMIT',
    ACCEPT = 'ACCEPT',
    DECLINE = 'DECLINE'
}